ref(server): Compile quotas only under processing
